Add a confirmation dialog to deletion.

+++ src/app/pages/playlist-editor/playlist-editor.ts
@@ -10,6 +10,8 @@ import { MatIconModule } from "@angular/material/icon";
 import { MatInputModule } from "@angular/material/input";
 import { MatSlideToggleModule } from "@angular/material/slide-toggle";
 import { PlaylistEntryListComponent } from "../../components/playlist-entry-list/playlist-entry-list.component";
+import { ConfirmationDialogComponent } from "../../components/confirmation-dialog/confirmation-dialog.component";
+import { MatDialog } from "@angular/material/dialog";
 
 @Component({
   selector: "app-playlist-editor",
@@ -20,6 +22,7 @@ import { PlaylistEntryListComponent } from "../../components/playlist-entry-list
     MatIconModule,
     MatInputModule,
     MatSlideToggleModule,
+    ConfirmationDialogComponent,
   ],
   providers: [QuotaService, DateService],
   templateUrl: "./playlist-editor.html",
@@ -29,6 +32,7 @@ export class PlaylistEditorPage implements OnInit {
   store = inject(LoggerStore);
   quotaService = inject(QuotaService);
   dateService = inject(DateService);
+  private dialog = inject(MatDialog);
   private route = inject(ActivatedRoute);
 
   readonly playlist = computed(() =>
@@ -69,10 +73,21 @@ export class PlaylistEditorPage implements OnInit {
   }
 
   onEntryDeleted(index) {
-    let atIndex = this.entries().find((e) => e.index === index);
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: {
+        title: "Delete",
+        message: "Are you sure you want remove this entry?",
+      },
+    });
 
-    if (atIndex) {
-      this.store.deletePlaylistEntry(atIndex.id);
-    }
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        let atIndex = this.entries().find((e) => e.index === index);
+
+        if (atIndex) {
+          this.store.deletePlaylistEntry(atIndex.id);
+        }
+      }
+    });
   }
 }
